fixing issue with keyboard docs shortcut when you just move the cursor around
